feat: Ajouter ccache comme lanceur de compilateur dans le fichier __init__.py; ajuster les seuils et les paramètres dans jeffres_audio.py; corriger le traitement audio dans audio_record.cpp; créer un nouveau fichier test_jeffres.py pour les exemples de microphone

+++ src/src/turtleaudio/src/audio_record.cpp
@@ -249,8 +249,8 @@ private:
 
     for (size_t i = 0; i < BUFFER_SIZE; ++i) 
     {
-      *dst_left++ = static_cast<int16_t>(*src++);
-      *dst_right++ = static_cast<int16_t>(*src++);
+      *dst_left++ = *src++;
+      *dst_right++ = *src++;
     }
 
     msg.header.frame_id = std::to_string(frame_count);
